+++ Net_Core/DataAccess/SQLServer/Base/ADBInteraction.cs
@@ -8,18 +8,23 @@ using System.Data.SqlClient;
 using System.Globalization;
 using System.Linq;
 using System.Reflection;
+using System.Threading.Tasks;
 
 namespace GrowthWare.DataAccess.SQLServer.Base;
 
 /// <summary>
 /// Performs all data store interaction to SQL Server.
 /// </summary>
-public abstract class AbstractDBInteraction : IDBInteraction, IDisposable
+public abstract class AbstractDBInteraction : IDBInteraction
 {
-    private string m_ConnectionString = string.Empty;
-    private bool m_DisposedValue;
-    internal Logger m_Logger = Logger.Instance();
 
+#region Member Fields
+    private string m_ConnectionString = string.Empty;
+    
+    internal Logger m_Logger = Logger.Instance();
+#endregion
+
+#region Public Properties
     /// <summary>
     /// Used for all methods to connect to the database.
     /// </summary>
@@ -31,9 +36,17 @@ public abstract class AbstractDBInteraction : IDBInteraction, IDisposable
         }
         set
         {
-            this.m_ConnectionString = value;
-        }
+            if (!string.IsNullOrWhiteSpace(value))
+            {
+                this.m_ConnectionString = value.Trim();
+            }
+            else
+            {
+                throw new DataAccessLayerException("The ConnectionString value cannot be null or blank!");
+            }
+        }        
     }
+#endregion
 
     /// <summary>
     /// Performs a bulk upload of IDatabaseFunctions objects into the
@@ -192,35 +205,40 @@ public abstract class AbstractDBInteraction : IDBInteraction, IDisposable
     }
 
     /// <summary>
-    /// Implements IDispose
+    /// Executes an asynchronous SQL command or stored procedure and returns the value of <paramref name="execution"/>.
     /// </summary>
-    /// <param name="disposing">Boolean</param>
-    /// <remarks></remarks>
-    protected virtual void Dispose(bool disposing)
+    /// <typeparam name="T">The type of the value returned by <paramref name="execution"/>.</typeparam>
+    /// <param name="commandText">The command text to execute.</param>
+    /// <param name="sqlParameters">The parameters to use when executing the command.</param>
+    /// <param name="forceCommandText">If true, the command text is executed as a SQL command. If false, the command text is executed as a stored procedure.</param>
+    /// <param name="execution">A function that takes a <see cref="SqlCommand"/> and returns a <see cref="Task{T}"/>.</param>
+    /// <returns>The result of <paramref name="execution"/>.</returns>
+    /// <exception cref="DataAccessLayerException">Thrown if an error occurs while executing the SQL command.</exception>
+    private async Task<T> executeAsync<T>(string commandText, SqlParameter[] sqlParameters, bool forceCommandText, Func<SqlCommand, Task<T>> execution)
     {
-        // Check to see if Dispose has already been called.
-        if (!m_DisposedValue)
+        if (string.IsNullOrWhiteSpace(m_ConnectionString))
         {
-            if (disposing)
-            {
-                // // Dispose managed resources if you have any.
-            }
-            // TODO: free unmanaged resources (unmanaged objects) and override Finalize() below.
-            // TODO: set large fields to null.
-
+            throw new DataAccessLayerException("The ConnectionString property cannot be null or blank!");
         }
-        m_DisposedValue = true;
-    }
-
-    /// <summary>
-    /// Implements Dispose
-    /// </summary>
-    /// <remarks></remarks>
-    public void Dispose()
-    {
-        //Do not change this code.  Put cleanup code in Dispose(bool disposing) above.
-        Dispose(true);
-        GC.SuppressFinalize(this);
+        await using var mSqlConnection = new SqlConnection(m_ConnectionString);
+        await mSqlConnection.OpenAsync();
+        await using var command = new SqlCommand(commandText, mSqlConnection)
+        {
+            CommandType = forceCommandText ? CommandType.Text : CommandType.StoredProcedure
+        };
+        if (sqlParameters != null)
+        {
+            command.Parameters.AddRange(sqlParameters);
+        }
+        try
+        {
+            return await execution(command);
+        }
+        catch (Exception ex)
+        {
+            m_Logger.Error(ex);
+            throw new DataAccessLayerException("Error executing SQL command.", ex);
+        }
     }
 
     /// <summary>
@@ -292,6 +310,51 @@ public abstract class AbstractDBInteraction : IDBInteraction, IDisposable
     }
 
     /// <summary>
+    /// Executes an asynchronous SQL command or stored procedure, and returns the number of rows affected.
+    /// </summary>
+    /// <param name="commandText">The command text to execute.</param>
+    /// <param name="sqlParameters">The parameters to use when executing the command.</param>
+    /// <param name="forceCommandText">If true, the command text is executed as a SQL command. If false, the command text is executed as a stored procedure.</param>
+    /// <returns>The number of rows affected by the execution of the SQL command.</returns>
+    /// <exception cref="DataAccessLayerException">Thrown if an error occurs while executing the SQL command.</exception>
+    protected virtual Task<int> ExecuteNonQueryAsync(string commandText, SqlParameter[] sqlParameters = null, bool forceCommandText = false)
+        => executeAsync(commandText, sqlParameters, forceCommandText, cmd => cmd.ExecuteNonQueryAsync());
+
+    /// <summary>
+    /// Executes a non query asynchronously given the commandText and sql parameters if any
+    /// </summary>
+    /// <param name="commandText">String</param>
+    /// <returns>int</returns>
+    /// <exception cref="DataAccessLayerException"></exception>
+    /// <remarks>This overload will assume the commandText is a SQL command and not a stored procedure.</remarks>
+    protected virtual Task<int> ExecuteNonQueryAsync(string commandText)
+    {
+        return this.ExecuteNonQueryAsync(commandText, null, true);
+    }
+
+    /// <summary>
+    /// Executes a scalar query asynchronously against the database.
+    /// </summary>
+    /// <param name="commandText">The command text to execute.</param>
+    /// <param name="sqlParameters">The parameters to use when executing the command.</param>
+    /// <param name="forceCommandText">If true, the command text is executed as a SQL command. If false, the command text is executed as a stored procedure.</param>
+    /// <returns>The value of the first column of the first row of the result set returned by the query. If the result set contains no rows, this method returns <see cref="DBNull.Value"/>.</returns>
+    /// <exception cref="DataAccessLayerException">Thrown if an error occurs while executing the SQL command.</exception>
+    protected virtual Task<object> ExecuteScalarAsync(string commandText, SqlParameter[] sqlParameters = null, bool forceCommandText = false)
+        => executeAsync(commandText, sqlParameters, forceCommandText, cmd => cmd.ExecuteScalarAsync());
+
+    /// <summary>
+    /// Executes a scalar query asynchronously against the database.
+    /// </summary>
+    /// <param name="commandText">The command text to execute.</param>
+    /// <returns>this.ExecuteScalarAsync(commandText, null, true)</returns>
+    /// <exception cref="DataAccessLayerException">Thrown if an error occurs while executing the SQL command.</exception>
+    protected virtual Task<object> ExecuteScalarAsync(string commandText)
+    {
+        return this.ExecuteScalarAsync(commandText, null, true);
+    }
+
+    /// <summary>
     /// Executes a scalar query against the database.
     /// This function takes a command text and optional SQL parameters, and returns the result of the query.
     /// If forceCommandText is set to false and SQL parameters are provided, the command type is set to StoredProcedure.
@@ -301,7 +364,7 @@ public abstract class AbstractDBInteraction : IDBInteraction, IDisposable
     /// <param name="sqlParameters">The SQL parameters to use with the command.</param>
     /// <param name="forceCommandText">Whether to force the command type to Text, even if SQL parameters are provided.</param>
     /// <returns></returns>
-    protected object ExecuteScalar(string commandText, SqlParameter[] sqlParameters, bool forceCommandText = false)
+    protected object ExecuteScalar(string commandText, SqlParameter[] sqlParameters = null, bool forceCommandText = false)
     {
         this.IsValid();
         try
@@ -477,6 +540,12 @@ public abstract class AbstractDBInteraction : IDBInteraction, IDisposable
         return this.GetDataRow(commandText, null);
     }
 
+    protected virtual async Task<DataRow> GetDataRowAsync(string commandText, SqlParameter[] sqlParameters = null, bool forceCommandText = false)
+    {
+        var mRetVal = await GetDataTableAsync(commandText, sqlParameters, forceCommandText);
+        return mRetVal.Rows.Count > 0 ? mRetVal.Rows[0] : null;
+    }
+
     /// <summary>
     /// Returns a DataSet given the store procedure and sql parameters
     /// </summary>
@@ -543,6 +612,18 @@ public abstract class AbstractDBInteraction : IDBInteraction, IDisposable
         return this.GetDataSet(commandText, null);
     }
 
+    protected virtual async Task<DataSet> GetDataSetAsync(string commandText, SqlParameter[] sqlParameters = null, bool forceCommandText = false)
+    {
+        var mRetVal = new DataSet();
+        await executeAsync(commandText, sqlParameters, forceCommandText, async mSqlCommand =>
+        {
+            using var mSqlDataAdapter = new SqlDataAdapter(mSqlCommand);
+            await Task.Run(() => mSqlDataAdapter.Fill(mRetVal));
+            return 0;
+        });
+        return mRetVal;
+    }
+
     /// <summary>
     /// Returns a DataTable given the command text, Sql Connection and sql parameters
     /// </summary>
@@ -566,6 +647,18 @@ public abstract class AbstractDBInteraction : IDBInteraction, IDisposable
     protected virtual DataTable GetDataTable(String commandText)
     {
         return this.GetDataTable(commandText, null);
+    }
+    
+    protected virtual async Task<DataTable> GetDataTableAsync(string commandText, SqlParameter[] sqlParameters = null, bool forceCommandText = false)
+    {
+        DataTable mRetVal = new();
+        await executeAsync(commandText, sqlParameters, forceCommandText, async mSqlCommand =>
+        {
+            await using var reader = await mSqlCommand.ExecuteReaderAsync();
+            mRetVal.Load(reader);
+            return 0;
+        });
+        return mRetVal;
     }
 
     /// <summary>
@@ -645,12 +738,5 @@ public abstract class AbstractDBInteraction : IDBInteraction, IDisposable
     {
         this.isConnectionStringSet();
     }
-    
-    //// TODO: override Finalize() only if Dispose(ByVal disposing As Boolean) above has code to free unmanaged resources.
-    //~DDBInteraction() 
-    //{
-    //    // Do not change this code.  Put cleanup code in Dispose(ByVal disposing As Boolean) above.
-    //    Dispose(false);
-    //}
 
 }